+++ tests/testthat/test_repos.R
@@ -9,40 +9,34 @@ my_project <- gl_project_connection(test_url,
 
 
 test_that("Repo access works", {
-  
-  expect_is(my_gitlab(gl_repository, project = 1), "data.frame")
+
   expect_is(my_gitlab(gl_repository, project = test_project), "data.frame")
   expect_is(my_gitlab(gl_repository, project = test_project, "contributors"), "data.frame")
   
   expect_is(my_gitlab(gl_list_files, test_project), "data.frame")
-  expect_is(my_gitlab(gl_get_file, test_project, "README.md", force_api_v3 = (test_api_version == "v3")), "character")
-  
-  ## same with function idiom
+  expect_is(my_gitlab(gl_get_file, test_project, "README.md"), "character")
 
-  expect_is(gl_repository(project = 1, gitlab_con = my_gitlab), "data.frame")
+  ## same with function idiom
   expect_is(gl_repository(project = test_project, gitlab_con = my_gitlab), "data.frame")
   expect_is(gl_repository("contributors", project = test_project, gitlab_con = my_gitlab), "data.frame")
   expect_is(gl_list_files(test_project, gitlab_con = my_gitlab), "data.frame")
-  expect_is(gl_get_file(test_project, "README.md", gitlab_con = my_gitlab, force_api_v3 = (test_api_version == "v3")), "character")
+  expect_is(gl_get_file(test_project, "README.md", gitlab_con = my_gitlab), "character")
   
   ## same with project connection
   
   expect_is(my_project(gl_repository), "data.frame")
   expect_is(my_project(gl_repository, "contributors"), "data.frame")
-  expect_is(my_project(gl_get_file, file_path = "README.md", force_api_v3 = (test_api_version == "v3")), "character")
+  expect_is(my_project(gl_get_file, file_path = "README.md"), "character")
 
   ## same with project connection & function idiom
   
   expect_is(gl_repository(gitlab_con = my_project), "data.frame")
   expect_is(gl_repository("contributors", gitlab_con = my_project), "data.frame")
-  expect_is(gl_get_file(file_path = "README.md", gitlab_con = my_project, force_api_v3 = (test_api_version == "v3")), "character")
+  expect_is(gl_get_file(file_path = "README.md", gitlab_con = my_project), "character")
   
   
   ## old API
-  expect_warning(my_gitlab(repository, project = 1), regexp = "deprecated")
-  
-  
-  
+  expect_warning(my_gitlab(repository, project = test_project), regexp = "deprecated")
 })
 
 test_that("Commits and diffs work", {