+++ R/gitlab_api.R
@@ -4,9 +4,6 @@
 #' use this function directly too often, but either use {gitlabr}'s convenience wrappers or write your
 #' own. See the {gitlabr} vignette for more information on this.
 #' 
-#' Note: currently GitLab API v3 is supported. Support for GitLab API v4 (for GitLab version >= 9.0) will
-#' be added soon.
-#' 
 #' @param req vector of characters that represents the call (e.g. `c("projects", project_id, "events")`)
 #' @param api_root URL where the GitLab API to request resides (e.g. `https://gitlab.myserver.com/api/v3/`)
 #' @param verb http verb to use for request in form of one of the `httr` functions
@@ -28,19 +25,50 @@
 #' @param argname_verb name of the argument of the verb that fields and information are passed on to
 #' @param ... named parameters to pass on to GitLab API (technically: modifies query parameters of request URL),
 #' may include private_token and all other parameters as documented for the GitLab API
+#' 
 #' @importFrom utils capture.output
 #' @importFrom tibble tibble as_tibble
 #' @export
 #' 
 #' @return the response from the GitLab API, usually as a `tibble` and including all pages
 #' 
+#' @details 
+#' `gitlab()` function allows to use any request of the GitLab API [https://docs.gitlab.com/ce/api/].
+#'
+#'  For instance, the API documentation shows how to create a new project in 
+#'  [https://docs.gitlab.com/ee/api/projects.html#create-project]: 
+#'  
+#'  - The verb is `POST`
+#'  - The request is `projects`
+#'  - Required attributes are `name` or `path` (if `name` not set)
+#'  - `default_branch` is an attribute that can be set if wanted
+#'  
+#'  The corresponding use of `gitlab()` is:  
+#'  
+#'  ```
+#'  gitlab(
+#'    req = "projects",
+#'    verb = httr::POST,
+#'    name = "toto",
+#'    default_branch = "main"
+#'  )
+#'  ```
+#'  
+#' Note: currently GitLab API v4 is supported. GitLab API v3 is no longer supported, but
+#' you can give it a try.
+#' 
 #' @examples \dontrun{
-#' gitlab(req = "projects",
-#'        api_root = "https://gitlab.example.com/api/v4/",
-#'        private_token = "123####89")
-#' gitlab(req = c("projects", 21, "issues"),
-#'        state = "closed",
-#'        gitlab_con = my_gitlab)
+#' # Connect as a fixed user to a GitLab instance
+#' set_gitlab_connection(
+#'   gitlab_url = "https://gitlab.com", 
+#'   private_token = Sys.getenv("GITLAB_COM_TOKEN")
+#' )
+#' 
+#' # Use a simple request
+#' gitlab(req = "projects")
+#' # Use a combined request with extra parameters
+#' gitlab(req = c("projects", 1234, "issues"),
+#'        state = "closed")
 #' }
 gitlab <- function(req,
                    api_root,